+++ src/solidmodels/postrender.jl
@@ -2,8 +2,7 @@
 function _postrender!(sm::SolidModel, operations)
     # Operations
     for (destination, op, args, kwargs...) in operations
-        dimtags = op(sm, args...; kwargs...)
-        sm[destination] = dimtags
+        sm[destination] = op(sm, args...; kwargs...)
     end
 end
 
@@ -504,7 +503,9 @@ end
 """
     restrict_to_volume(sm::SolidModel, volume)
 
-Replace all entities and groups with their intersection with `sm[volume, 3]`.
+Checks if all surfaces and volumes are contained within `sm[volume, 3]`, and if not performs
+an intersection operation replacing all entities and groups with their intersection with
+`sm[volume, 3]`.
 
 Embeds entities if they are on the boundary of higher-dimensional entities and removes
 duplicate entities.
@@ -513,7 +514,23 @@ Preserves the meaning of existing groups by assigning to them the (possibly new)
 corresponding to that group's intersection with the volume.
 """
 function restrict_to_volume!(sm::SolidModel, volume)
-    dims = [3, 2, 1, 0]
+
+    # Check if the subtraction of the bounding volume from all surfaces and volumes is the
+    # empty set.
+    dims = SVector(3, 2, 1)
+    groups =
+        [(name, dimtags(pg)) for dim in dims for (name, pg) in pairs(dimgroupdict(sm, dim))]
+    allents = vcat([gmsh.model.get_entities(dim) for dim in dims]...)
+
+    out_dim_tags, _ = kernel(sm).cut(allents, dimtags(sm[volume, 3]), -1, false, false)
+    isempty(out_dim_tags) && return dimtags(sm[volume, 3])
+
+    # There were entities found after cutting, the restricting volume is a subset of the
+    # rendered geometry, will need to perform the intersection.
+    kernel(sm).remove(out_dim_tags, true)
+    _synchronize!(sm)
+
+    dims = SVector(3, 2, 1, 0)
     groups =
         [(name, dimtags(pg)) for dim in dims for (name, pg) in pairs(dimgroupdict(sm, dim))]
     allents = vcat([gmsh.model.get_entities(dim) for dim in dims]...)